Renaming some files. Re-create a new thread and fix: raise RuntimeError("threads can only be started once")
